requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
amqp==5.3.1
    # via kombu
asgiref==3.11.1
    # via django
billiard==4.2.4
    # via celery
celery==5.4.0
    # via
    #   -r requirements/base.in
    #   django-celery-beat
click==8.3.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
django-colorfield==0.14.0
    # via django-admin-interface
cron-descriptor==2.0.6
    # via django-celery-beat
django==5.1.15
    # via
    #   -r requirements/base.in
    #   django-celery-beat
    #   django-cors-headers
    #   django-filter
    #   django-redis
    #   django-timezone-field
    #   djangorestframework
django-admin-interface==0.26.0

django-celery-beat==2.8.1
    # via -r requirements/base.in
django-cors-headers==4.3.0
    # via -r requirements/base.in
django-filter==25.1
    # via -r requirements/base.in
django-redis==6.0.0
    # via -r requirements/base.in
django-timezone-field==7.2.1
    # via django-celery-beat
djangorestframework==3.16.1
    # via -r requirements/base.in
kombu==5.6.2
    # via celery
packaging==26.0
    # via kombu
phonenumbers==9.0.23
    # via -r requirements/base.in
pillow==12.1.0
    # via -r requirements/base.in
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.11
    # via -r requirements/base.in
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.in
    #   celery
python-decouple==3.8
    # via -r requirements/base.in
python-json-logger==2.0.7
    # via -r requirements/base.in
pyjwt==2.8.0
    # via -r requirements/base.in
cryptography==41.0.7
    # via -r requirements/base.in
requests==2.31.0
    # via -r requirements/base.in
svix==1.14.0
    # via -r requirements/base.in
redis==7.1.0
    # via
    #   -r requirements/base.in
    #   django-redis
reportlab==4.0.7
    # via -r requirements/base.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.5
    # via django
typing-extensions==4.15.0
    # via cron-descriptor
tzdata==2025.3
    # via
    #   celery
    #   django-celery-beat
    #   kombu
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.6.0
    # via prompt-toolkit
firebase-admin==6.2.0

# Channels
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
liqpay-python==1.0
    # via -r requirements/base.in
